make t-edge linear
